NPC and Cast handler renaming
