+++ EOLib/PacketHandlers/RefreshMapStateHandler.cs
@@ -3,12 +3,13 @@ using EOLib.Domain.Character;
 using EOLib.Domain.Login;
 using EOLib.Domain.Map;
 using EOLib.Domain.Notifiers;
-using EOLib.Domain.NPC;
 using EOLib.Net;
 using EOLib.Net.Handlers;
 using EOLib.Net.Translators;
 using System.Collections.Generic;
 using System.Linq;
+
+using DomainNPC = EOLib.Domain.NPC.NPC;
 
 namespace EOLib.PacketHandlers
 {
@@ -54,7 +55,7 @@ namespace EOLib.PacketHandlers
                 .WithRenderProperties(updatedRenderProperties);
 
             _currentMapStateRepository.Characters = data.Characters.ToDictionary(k => k.ID, v => v);
-            _currentMapStateRepository.NPCs = new HashSet<NPC>(data.NPCs);
+            _currentMapStateRepository.NPCs = new HashSet<DomainNPC>(data.NPCs);
             _currentMapStateRepository.MapItems = new HashSet<MapItem>(data.Items);
 
             _currentMapStateRepository.OpenDoors.Clear();
